+++ ksp/src/jvmMain/kotlin/eu/vendeli/ksp/ActivityCollectors.kt
@@ -11,6 +11,7 @@ import com.squareup.kotlinpoet.STRING
 import com.squareup.kotlinpoet.TypeVariableName
 import com.squareup.kotlinpoet.asTypeName
 import com.squareup.kotlinpoet.buildCodeBlock
+import eu.vendeli.ksp.dto.CallbackQueryParams
 import eu.vendeli.ksp.dto.CollectorsContext
 import eu.vendeli.ksp.dto.CommonAnnotationData
 import eu.vendeli.ksp.utils.addMap
@@ -63,6 +64,7 @@ internal fun collectCommandActivities(
         val guardAnnotationData = function.parseAnnotatedGuard()
         val rateLimitsAnnotationData = function.parseAnnotatedRateLimits()
         val argParserAnnotationData = function.parseAnnotatedArgParser()
+        val params = if (annotationData.isAutoAnswer) listOf(CallbackQueryParams.AutoAnswer) else emptyList()
 
         annotationData.value.forEach {
             annotationData.scope.forEach { updT ->
@@ -82,6 +84,7 @@ internal fun collectCommandActivities(
                             guardClass = guardAnnotationData ?: annotationData.guardClass,
                             argParserClass = argParserAnnotationData ?: annotationData.argParserClass,
                         ),
+                        params
                     ),
                 )
             }