+++ ksp/src/jvmMain/kotlin/eu/vendeli/ksp/InvocationLambdaBuilder.kt
@@ -14,6 +14,8 @@ import com.squareup.kotlinpoet.STRING
 import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.buildCodeBlock
 import com.squareup.kotlinpoet.ksp.toTypeName
+import eu.vendeli.ksp.dto.CallbackQueryParams
+import eu.vendeli.ksp.dto.LambdaParameters
 import eu.vendeli.ksp.utils.FileBuilder
 import eu.vendeli.ksp.utils.botClass
 import eu.vendeli.ksp.utils.businessConnectionUpdateClass
@@ -78,6 +80,7 @@ internal fun FileBuilder.buildInvocationLambdaCodeBlock(
     injectableTypes: Map<TypeName, ClassName>,
     pkg: String? = null,
     meta: Pair<String, Array<Any?>>? = null,
+    parameters: List<LambdaParameters> = emptyList()
 ) = buildCodeBlock {
     val isTopLvl = function.functionKind == FunctionKind.TOP_LEVEL
     val funQualifier = function.qualifiedName!!.getQualifier()
@@ -180,6 +183,14 @@ internal fun FileBuilder.buildInvocationLambdaCodeBlock(
                 funQualifier,
                 pkg,
             )
+
+            if (parameters.contains(CallbackQueryParams.AutoAnswer)) {
+                addImport("eu.vendeli.tgbot.api.answer", "answerCallbackQuery")
+                addImport("eu.vendeli.tgbot.types.internal", "CallbackQueryUpdate")
+                addImport("eu.vendeli.tgbot.types.internal", "getUser")
+                add("answerCallbackQuery((update as CallbackQueryUpdate).callbackQuery.id).send(update.getUser(), bot)")
+            }
+
             add("\n%L.invoke(\n\t%L\n)\n", funName, parametersEnumeration)
         }.endControlFlow()
         .build()