+++ ksp/src/jvmMain/kotlin/eu/vendeli/ksp/dto/AnnotationData.kt
@@ -8,4 +8,5 @@ internal data class AnnotationData(
     val scope: List<UpdateType>,
     val guardClass: String,
     val argParserClass: String,
+    val isAutoAnswer: Boolean
 )