+++ ksp/src/jvmMain/kotlin/eu/vendeli/ksp/dto/LambdaParameters.kt
@@ -1,0 +1,8 @@
+package eu.vendeli.ksp.dto
+
+sealed class LambdaParameters
+
+sealed class CallbackQueryParams : LambdaParameters() {
+    data object AutoAnswer : CallbackQueryParams()
+}
+